Configure .ruby-gemset to scope project gems independently
Did this to take full advantage of RVM in addition to the project
already having .ruby-version. Also, that makes it possible to configure
same gem versions for different projects, but with different native
binding configurations or CPU architectures (e.g. x86 vs x64). Also,
it makes it easier to delete all project gems from machine when
removing project or upgrading to a new version of Ruby, without
distrubing other project gems that might be dependent on older
version of Ruby.